Shuffled the review data cache out to its own class.
